Refactor facades into their own files
